Make deregistration_delay attribute customizable for ALB Target Groups (#42)
* Make deregistration_delay attribute customizable for ALB Target Groups

The `deregistration_delay` attribute of a Target Group is set to 300 seconds by default
as a conservatively safe value. However, if a service owner knows that their connections stop faster
it makes sense to reduce the delay to speed up the deployments. This PR makes the customization
possible via the module.

* Ensure README.md and CHANGELOG.md are up to date